api_spotify/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
db-dtypes==1.0.2
google-api-core==2.8.2
google-auth==2.8.0
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
oni-notifications-helper==1.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.6
proto-plus==1.20.6
protobuf==4.21.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
